Move show-checks flag to main.rs
