+++ src/lib.rs
@@ -1,74 +1,19 @@
 use crate::common::*;
 
-use clap::Arg;
 use std::error::Error;
-use std::ffi::OsStr;
 use std::path::PathBuf;
-use std::{env, process};
 
 mod checks;
 mod common;
 mod fs_utils;
 
-fn get_args(current_dir: &OsStr) -> clap::ArgMatches {
-    clap::App::new(env!("CARGO_PKG_NAME"))
-        .about(env!("CARGO_PKG_DESCRIPTION"))
-        .author(env!("CARGO_PKG_AUTHORS"))
-        .version(env!("CARGO_PKG_VERSION"))
-        .version_short("v")
-        .arg(
-            Arg::with_name("input")
-                .help("files or paths")
-                .index(1)
-                .default_value_os(current_dir)
-                .required(true)
-                .multiple(true),
-        )
-        .arg(
-            Arg::with_name("exclude")
-                .short("e")
-                .long("exclude")
-                .value_name("FILE_NAME")
-                .help("Excludes files from check")
-                .multiple(true)
-                .takes_value(true),
-        )
-        .arg(
-            Arg::with_name("skip")
-                .short("s")
-                .long("skip")
-                .value_name("CHECK_NAME")
-                .help("Skips checks")
-                .multiple(true)
-                .takes_value(true),
-        )
-        .arg(
-            Arg::with_name("show-checks")
-                .long("show-checks")
-                .help("Shows list of available checks"),
-        )
-        .get_matches()
-}
+pub use checks::available_check_names;
 
 #[allow(clippy::redundant_closure)]
-pub fn run() -> Result<Vec<Warning>, Box<dyn Error>> {
-    let current_dir = match env::current_dir() {
-        Ok(dir) => dir,
-        Err(e) => return Err(Box::new(e)),
-    };
-
-    let args = get_args(current_dir.as_os_str());
-
+pub fn run(args: clap::ArgMatches, current_dir: &PathBuf) -> Result<Vec<Warning>, Box<dyn Error>> {
     let mut dirs: Vec<PathBuf> = Vec::new();
     let mut file_paths: Vec<PathBuf> = Vec::new();
     let mut skip_checks: Vec<&str> = Vec::new();
-
-    if args.is_present("show-checks") {
-        checks::available_check_names()
-            .iter()
-            .for_each(|name| println!("{}", name));
-        process::exit(0);
-    }
 
     if let Some(skip) = args.values_of("skip") {
         skip_checks = skip.collect();
@@ -121,7 +66,7 @@ pub fn run() -> Result<Vec<Warning>, Box<dyn Error>> {
 
         let file_with_lines = match FileEntry::from(relative_path) {
             Some(f) => f,
-            _ => continue,
+            None => continue,
         };
 
         let lines = get_line_entries(file_with_lines.0, file_with_lines.1);
